requirements.txt
langchain
langchain-chroma
langchain-huggingface
langchain-community
transformers
accelerate
chromadb
langchain-core
langchain-text-splitters
requirements-parser
neo4j
langchain-experimental
streamlit
watchdog